refactor(module): add type template
