+++ src/kit.ts
@@ -1,36 +1,8 @@
 import type { Nuxt } from '@nuxt/schema'
 import type { NitroConfig } from 'nitropack'
 import type { NuxtModule } from 'nuxt/schema'
-import { addTemplate, createResolver, loadNuxtModuleInstance, useNuxt } from '@nuxt/kit'
-import { relative } from 'pathe'
+import { loadNuxtModuleInstance, useNuxt } from '@nuxt/kit'
 import { env, provider } from 'std-env'
-
-export function extendTypes(module: string, template: (options: { typesPath: string }) => string | Promise<string>) {
-  const nuxt = useNuxt()
-  const { resolve } = createResolver(import.meta.url)
-  // paths.d.ts
-  addTemplate({
-    filename: `module/${module}.d.ts`,
-    getContents: async () => {
-      const typesPath = relative(resolve(nuxt!.options.rootDir, nuxt!.options.buildDir, 'module'), resolve('runtime/types'))
-      const s = await template({ typesPath })
-      return `// Generated by ${module}
-${s}
-export {}
-`
-    },
-  })
-
-  nuxt.hooks.hook('prepare:types', ({ references }) => {
-    references.push({ path: resolve(nuxt.options.buildDir, `module/${module}.d.ts`) })
-  })
-  nuxt.hooks.hook('nitro:config', (config) => {
-    config.typescript = config.typescript || {}
-    config.typescript.tsConfig = config.typescript.tsConfig || {}
-    config.typescript.tsConfig.include = config.typescript.tsConfig.include || []
-    config.typescript.tsConfig.include.push(`./module/${module}.d.ts`)
-  })
-}
 
 const autodetectableProviders = {
   azure_static: 'azure',
