+++ src/module.ts
@@ -6,6 +6,7 @@ import {
   addServerHandler,
   addServerImportsDir,
   addServerPlugin,
+  addTypeTemplate,
   createResolver,
   defineNuxtModule,
   hasNuxtModule,
@@ -19,11 +20,13 @@ import { withoutTrailingSlash, withTrailingSlash } from 'ufo'
 import { AiBots, NonHelpfulBots } from './const'
 import { setupDevToolsUI } from './devtools'
 import { resolveI18nConfig, splitPathForI18nLocales } from './i18n'
-import { extendTypes, isNuxtGenerate, resolveNitroPreset } from './kit'
+import { isNuxtGenerate, resolveNitroPreset } from './kit'
 import { logger } from './logger'
 import {
-  asArray,
   normaliseRobotsRouteRule,
+} from './runtime/server/nitro'
+import {
+  asArray,
   normalizeGroup,
   parseRobotsTxt,
   validateRobots,
@@ -461,8 +464,14 @@ export default defineNuxtModule<ModuleOptions>({
       }
     })
 
-    extendTypes('nuxt-robots', ({ typesPath }) => {
-      return `
+    addTypeTemplate({
+      filename: 'module/nuxt-robots.d.ts',
+      getContents: (data) => {
+        const typesPath = relative(resolve(data.nuxt!.options.rootDir, data.nuxt!.options.buildDir, 'module'), resolve('runtime/types'))
+        return `// Generated by nuxt-robots
+
+import type { RobotsContext } from '#robots/types'
+
 declare module 'nitropack/types' {
   interface NitroApp {
     _robots: {
@@ -488,6 +497,7 @@ declare module 'nitropack/types' {
     'robots:robots-txt': (ctx: import('${typesPath}').HookRobotsTxtContext) => void | Promise<void>
   }
 }
+
 declare module 'nitropack' {
   interface NitroApp {
     _robots: {
@@ -513,13 +523,18 @@ declare module 'nitropack' {
     'robots:robots-txt': (ctx: import('${typesPath}').HookRobotsTxtContext) => void | Promise<void>
   }
 }
+
 declare module 'h3' {
-  import type { RobotsContext } from '#robots/types'
   interface H3EventContext {
     robots: RobotsContext
   }
 }
+
+export {}
 `
+      },
+    }, {
+      nitro: true,
     })
 
     // only prerender for `nuxi generate`
